requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.26.89
botocore==1.29.89
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
Django==4.1.7
django-cors-headers==3.13.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
jsonfield==3.1.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
zope.interface==5.5.2
